Updates & fixes * Minor fixes in API Toolkit. * Updated API Simulator to closely represent the actual API. * Replaced menu buttons to back buttons in screens that Drawer is not necessary. * Disable drawer navigation in screens that are not necessary.
