+++ src/screens/SampleTagging/Analysis/Analysis.js
@@ -50,8 +50,8 @@ class Analysis extends React.Component {
       <Container>
         <Header>
           <Left>
-            <Button transparent icon onPress={() => navigation.toggleDrawer()}>
-              <Icon type="MaterialCommunityIcons" name="menu" />
+            <Button transparent icon onPress={() => navigation.goBack()}>
+              <Icon type="MaterialCommunityIcons" name="arrow-left" />
             </Button>
           </Left>
           <Body>
