+++ src/screens/SampleTagging/router.js
@@ -4,7 +4,7 @@ import Analysis from './Analysis/Analysis'
 import Search from './Analysis/Search'
 import Tagging from './Analysis/Tagging'
 
-export default createStackNavigator(
+const sampleTaggingRouter = createStackNavigator(
   {
     recentScans: RecentScans,
     search: Search,
@@ -13,3 +13,11 @@ export default createStackNavigator(
   },
   { defaultNavigationOptions: { header: null } }
 )
+
+sampleTaggingRouter.navigationOptions = ({ navigation }) => {
+  return {
+    drawerLockMode: (navigation.state.index > 1) ? 'locked-closed' : 'unlocked'
+  }
+}
+
+export default sampleTaggingRouter
